Small changes... Refactoring, descriptions, packages names...
