+++ books/pom.xml
@@ -9,10 +9,10 @@
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.rho</groupId>
-	<artifactId>books</artifactId>
+	<artifactId>books-service</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>books</name>
-	<description>Eureka client</description>
+	<name>Book Service</name>
+	<description>Eureka client that presents list of books</description>
 
 	<properties>
 		<java.version>1.8</java.version>
