+++ books/src/main/resources/application.properties
@@ -2,11 +2,3 @@
 spring.application.name=book-service
 server.port=8200
 eureka.client.serviceUrl.defaultZone= http://localhost:8080/eureka/
-
-#Eureka server url for registering
-#This is eureka client
-#eureka.client.register-with-eureka=true
-#eureka.client.fetch-registry=true
-#register url to server
-#eureka.client.serviceUrl.defaultZone=http://localhost:8080/eureka/
-#instance.hostname=localhost
