+++ series-service/pom.xml
@@ -8,11 +8,11 @@
 		<version>2.1.6.RELEASE</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>com.example</groupId>
+	<groupId>com.rho</groupId>
 	<artifactId>series-service</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>series-service</name>
-	<description>Netflix Series Service</description>
+	<description>Eureka client that presents list of netflix series</description>
 
 	<properties>
 		<java.version>1.8</java.version>
