+++ series-service/src/main/java/com/rho/series/SeriesServiceApplication.java
@@ -1,17 +1,15 @@
-package com.example.zuulserver;
+package com.rho.series;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
 
 @SpringBootApplication
 @EnableDiscoveryClient
-@EnableZuulProxy
-public class ZuulServerApplication {
+public class SeriesServiceApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(ZuulServerApplication.class, args);
+		SpringApplication.run(SeriesServiceApplication.class, args);
 	}
 
 }
