+++ series-service/src/main/java/com/rho/series/service/ISerieService.java
@@ -1,6 +1,6 @@
-package com.example.series.service;
+package com.rho.series.service;
 
-import com.example.series.bean.NetflixSerie;
+import com.rho.series.bean.NetflixSerie;
 
 import java.util.List;
