+++ series-service/src/test/java/com/rho/series/SeriesServiceApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.zuulserver;
+package com.rho.series;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -7,7 +7,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
-public class ZuulServerApplicationTests {
+public class SeriesServiceApplicationTests {
 
 	@Test
 	public void contextLoads() {
