+++ zuul-server/src/main/java/com/rho/zuulserver/ZuulServerApplication.java
@@ -1,15 +1,17 @@
-package com.example.series;
+package com.rho.zuulserver;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
 
 @SpringBootApplication
 @EnableDiscoveryClient
-public class SeriesServiceApplication {
+@EnableZuulProxy
+public class ZuulServerApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(SeriesServiceApplication.class, args);
+		SpringApplication.run(ZuulServerApplication.class, args);
 	}
 
 }
