+++ zuul-server/src/test/java/com/rho/zuulserver/ZuulServerApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.series;
+package com.rho.zuulserver;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -7,7 +7,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
-public class SeriesServiceApplicationTests {
+public class ZuulServerApplicationTests {
 
 	@Test
 	public void contextLoads() {
